requirements.txt
boto3==1.5.22
botocore==1.8.36
cffi==1.11.4
characteristic==14.3.0
cryptography==2.1.4
cssselect==1.0.3
dj-database-url==0.4.2
Django==2.0.1
django-mailgun==0.9.1
docutils==0.14
enum34==1.1.6
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.19
jmespath==0.9.3
lxml==4.1.1
Pillow==5.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyOpenSSL==17.5.0
python-dateutil==2.6.1
pytz==2017.3
queuelib==1.4.2
requests==2.18.4
Scrapy==1.5.0
scrapy-djangoitem==1.1.1
service-identity==17.0.0
six==1.11.0
sqlparse==0.2.4
titlecase==0.12.0
Twisted==17.9.0
w3lib==1.19.0
whitenoise==3.3.1
zope.interface==4.4.3
psycopg2==2.7.3.2
